feat: fix queues, better static server
