+++ .gitignore
@@ -22,6 +22,7 @@ var/
 wheels/
 share/python-wheels/
 *.egg-info/
+videos
 .installed.cfg
 *.egg
 MANIFEST
